Big refactor safety commit
