--- src/components/PivotTable/PivotTableRenderer/classes/GroupResult.ts
@@ -1,7 +1,0 @@
-export class GroupResult {
-  nodeValue: number
-
-  constructor(value: number) {
-    this.nodeValue = value
-  }
-}
